Fix inheritance browser indentions by removing overridden MessageSet >> #initializeMessageList:
The old version of #initializeMessageList: stored in this override is no longer compatible with modern message set indentions as used by the inheritance browser. The sole purpose of this override was to preserve CSMethodReferences while string elements from the array are parsed. Since eem 12/16/2016 11:25 (Squeak 5.2), this invariant is already satisifed in the base system, thus the extension method is not needed any longer.